move determineProminentWordsSpec files
