+++ packages/yoastseo/spec/languageProcessing/helpers/prominentWords/determineProminentWordsGermanSpec.js
@@ -1,8 +1,11 @@
-import ProminentWord from "../../src/values/ProminentWord";
-import { getProminentWords } from "../../src/languageProcessing/helpers/prominentWords/determineProminentWords";
-import getMorphologyData from "../specHelpers/getMorphologyData";
+import ProminentWord from "../../../../src/values/ProminentWord";
+import { getProminentWords } from "../../../../src/languageProcessing/helpers/prominentWords/determineProminentWords";
+import getMorphologyData from "../../../specHelpers/getMorphologyData";
+import Researcher from "../../../../src/languageProcessing/languages/de/Researcher";
 
-const morphologyData = getMorphologyData( "de" ).de;
+const morphologyData = getMorphologyData( "de" );
+const researcher = new Researcher();
+researcher.addResearchData( "morphology", morphologyData );
 
 describe( "gets German prominent words", function() {
 	it( "returns prominent words", function() {
@@ -23,7 +26,8 @@ describe( "gets German prominent words", function() {
 			new ProminentWord( "studieren", "studium", 48 ),
 		];
 
-		const words = getProminentWords( input, [], "de", morphologyData );
+		const words = getProminentWords( input, [], researcher.getHelper( "getStemmer" )( researcher ),
+			researcher.getConfig( "functionWords" ) );
 
 		expect( words ).toEqual( expected );
 	} );
