+++ packages/yoastseo/spec/languageProcessing/helpers/prominentWords/determineProminentWordsItalianSpec.js
@@ -1,5 +1,8 @@
-import ProminentWord from "../../src/values/ProminentWord";
-import { getProminentWords } from "../../src/languageProcessing/helpers/prominentWords/determineProminentWords";
+import ProminentWord from "../../../../src/values/ProminentWord";
+import { getProminentWords } from "../../../../src/languageProcessing/helpers/prominentWords/determineProminentWords";
+import Researcher from "../../../../src/languageProcessing/languages/it/Researcher";
+
+const researcher = new Researcher();
 
 describe( "gets Italian prominent words", function() {
 	it( "returns prominent words", function() {
@@ -21,7 +24,8 @@ describe( "gets Italian prominent words", function() {
 			new ProminentWord( "procura", "procura", 11 ),
 		];
 
-		const words = getProminentWords( input, [], "it", false );
+		const words = getProminentWords( input, [], researcher.getHelper( "getStemmer" )( researcher ),
+			researcher.getConfig( "functionWords" ) );
 
 		expect( words ).toEqual( expected );
 	} );
