+++ packages/yoastseo/spec/languageProcessing/helpers/prominentWords/determineProminentWordsPortugueseSpec.js
@@ -1,5 +1,8 @@
-import ProminentWord from "../../src/values/ProminentWord";
-import { getProminentWords } from "../../src/languageProcessing/helpers/prominentWords/determineProminentWords";
+import ProminentWord from "../../../../src/values/ProminentWord";
+import { getProminentWords } from "../../../../src/languageProcessing/helpers/prominentWords/determineProminentWords";
+import Researcher from "../../../../src/languageProcessing/languages/pt/Researcher";
+
+const researcher = new Researcher();
 
 describe( "gets Portuguese prominent words", function() {
 	it( "returns prominent words", function() {
@@ -31,7 +34,8 @@ describe( "gets Portuguese prominent words", function() {
 			new ProminentWord( "vítimas", "vítimas", 8 ),
 		];
 
-		const words = getProminentWords( input, [], "pt", false );
+		const words = getProminentWords( input, [], researcher.getHelper( "getStemmer" )( researcher ),
+			researcher.getConfig( "functionWords" ) );
 
 		expect( words ).toEqual( expected );
 	} );
